mv storage and tensor attribute
