+++ scripts/validate_api/validate_includes.py
@@ -43,6 +43,8 @@ EXPERIMENTAL_TENSOR_FORWARD_HEADERS = {
     "tt-metalium/experimental/tensor/spec/memory_config/memory_config.hpp": "ttnn/tensor/memory_config/memory_config.hpp",
     "tt-metalium/experimental/tensor/topology/distributed_tensor_configs.hpp": "ttnn/distributed/distributed_configs.hpp",
     "tt-metalium/experimental/tensor/topology/tensor_topology.hpp": "ttnn/distributed/tensor_topology.hpp",
+    "tt-metalium/experimental/tensor/details/storage.hpp": "ttnn/tensor/storage.hpp",
+    "tt-metalium/experimental/tensor/details/tensor_attributes.hpp": "ttnn/tensor/tensor_attributes.hpp",
 }
 
 # Files excluded from tensor forward header check - these are the forward headers themselves.
