+++ ttnn/CMakeLists.txt
@@ -148,9 +148,7 @@ target_sources(
         core/tensor/flatbuffer/tensor_spec_flatbuffer.cpp
         core/tensor/host_buffer/functions.cpp
         core/tensor/serialization.cpp
-        core/tensor/storage.cpp
         core/tensor/tensor.cpp
-        core/tensor/tensor_attributes.cpp
         core/tensor/tensor_impl.cpp
         core/tensor/tensor_ops.cpp
         core/tensor/tensor_utils.cpp
