+++ ttnn/api/ttnn/tensor/tensor_attributes.hpp
@@ -1,37 +1,8 @@
-// SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
+// SPDX-FileCopyrightText: © 2026 Tenstorrent AI ULC
 //
 // SPDX-License-Identifier: Apache-2.0
 
 #pragma once
 
-#include <memory>
-
-#include "ttnn/tensor/storage.hpp"
-#include "ttnn/tensor/tensor_spec.hpp"
-#include "ttnn/distributed/tensor_topology.hpp"
-
-namespace tt::tt_metal {
-
-class TensorAttributes : public std::enable_shared_from_this<TensorAttributes> {
-public:
-    TensorAttributes(Storage storage, TensorSpec tensor_spec, TensorTopology tensor_topology);
-    TensorAttributes(const TensorAttributes&) = default;
-    TensorAttributes(TensorAttributes&&) = default;
-    TensorAttributes& operator=(const TensorAttributes&) = default;
-    TensorAttributes& operator=(TensorAttributes&&) = default;
-
-    // Getters and setters.
-    const Storage& get_storage() const;
-    Storage& get_storage();
-    const TensorSpec& get_tensor_spec() const;
-    const TensorTopology& get_tensor_topology() const;
-
-    TensorAttributes with_tensor_topology(TensorTopology tensor_topology) const;
-
-private:
-    Storage storage_;
-    TensorSpec tensor_spec_;
-    TensorTopology tensor_topology_;
-};
-
-}  // namespace tt::tt_metal
+// Forward header - actual implementation is in tt-metalium experimental tensor.
+#include <tt-metalium/experimental/tensor/details/tensor_attributes.hpp>
